+++ module/timer/src/mod_timer.c
@@ -669,7 +669,7 @@ static int timer_start(fwk_id_t id)
 /* Module descriptor */
 const struct fwk_module module_timer = {
     .name = "Timer HAL",
-    .api_count = MOD_TIMER_API_COUNT,
+    .api_count = (unsigned int)MOD_TIMER_API_COUNT,
     .type = FWK_MODULE_TYPE_HAL,
     .init = timer_init,
     .element_init = timer_device_init,
